+++ firestore-pdf-generator/functions/lib/utilities/parameters.ts
@@ -6,7 +6,8 @@ export interface ParsedParameters {
   chromiumPdfOptions: PDFOptions;
   data: ParsedQs;
   headless: boolean;
-  outputBucketName: string | undefined;
+  outputStorageBucket: string | undefined;
+  outputStoragePrefix: string | undefined;
   outputFileName: string;
   templateBucket: string;
   templatePrefix: string;