+++ scripts/sync-lib.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+set -e
+
+for EXTENSION in "firestore-pdf-generator" "http-pdf-generator"
+do
+  rsync -av --delete common-stuff/lib/ $EXTENSION/functions/lib
+done